Porting from mock to actual rest api

Closes #61

See merge request cse1105/2020-2021/team-repositories/oopp-group-54/repository-template!36

+++ server/src/main/java/nl/tudelft/oopp/group54/controllers/questions/QuestionService.java
@@ -11,5 +11,5 @@ import java.util.List;
 public interface QuestionService {
     Map<String, Object> postQuestion(Integer lectureId, String userId, String questionText);
 
-    List<Question> getAllQuestions(Long lectureId, String userId);
+    List<Question> getAllQuestions(Integer lectureId, String userId);
 }
